backend/requirements.txt
flask==3.1.0
gunicorn==23.0.0
gevent==24.11.1
boto3==1.36.14
python-ulid==3.0.0
